requirements.txt
streamlit
requests
pandas
psycopg2-binary